Try to fix a tuning issue in si570_freqdial
In si570 driver's set_freq method, the driver distinguishes between a
small and large change in frequency with tuning being carried out
differently in the two cases. But if a large change in frequency is
carried out in small steps, the implementation is suspected to lead
to erroneous tuning. In this commit, a conservative approach to
retuning is taken in the si570_freqdial to possibly avoid the bug with
the driver as-is.